[otbn] Separate ELF-parsing code from simulator.
Put ELF-parsing code into otbn/util/shared so it can be used by other
tools.
